move active view state to local state
