+++ src/sidebar/Menu/Menu.tsx
@@ -9,21 +9,22 @@ import sessionSlice, { MenuContext, MenuType } from '../../store/slices/session'
 import { UnreachableCaseError } from '../../utils/UnreachableCaseError';
 import ContextMenu from './ContextMenu';
 import MoreMenu from './MoreMenu';
+import { View } from '../App';
 
-const renderMenu = (context: MenuContext) => {
+const renderMenu = (context: MenuContext, changeView: (value: View) => void) => {
     switch (context.type) {
         case MenuType.contextMenu:
             return <ContextMenu anchorPoint={context.anchorPoint} />;
 
         case MenuType.moreMenu:
-            return <MoreMenu anchorPoint={context.anchorPoint} />;
+            return <MoreMenu anchorPoint={context.anchorPoint} changeView={changeView} />;
 
         default:
             throw new UnreachableCaseError(context.type);
     }
 };
 
-const Menu: FunctionComponent = () => {
+const Menu = ({ changeView }: { changeView: (value: View) => void }) => {
     const context = useAppSelector((state) => state.session.menuContext);
     const visible = useAppSelector((state) => state.session.menuVisible);
 
@@ -37,9 +38,10 @@ const Menu: FunctionComponent = () => {
         return null;
     }
 
+    // TODO mr refactor the whole menu thing to not use redux
     return (
         <MenuBackdrop visible={visible} onMouseDown={hideMenu} onContextMenu={(e) => e.preventDefault()}>
-            {renderMenu(context)}
+            {renderMenu(context, changeView)}
         </MenuBackdrop>
     );
 };
