+++ src/store/slices/session.ts
@@ -20,14 +20,7 @@ export interface MenuContext {
     anchorPoint: Point;
 }
 
-export const enum View {
-    default = 'default', // needed to switch view on back button after background script was terminated
-    feedList = 'feedList',
-    subscribe = 'subscribe',
-}
-
 type SessionSliceState = {
-    activeView: View;
     // feedStatus is managed separately because it can contain entries that have no corresponding feed in feedSlice
     // (recently added and not successfully fetched or parsed)
     feedStatus: ReadonlyArray<{ url: string; status: FeedFetchStatus }>;
@@ -39,7 +32,6 @@ type SessionSliceState = {
 };
 
 export const initialState: SessionSliceState = {
-    activeView: View.default,
     feedStatus: [],
 
     menuContext: undefined,
@@ -54,10 +46,6 @@ const sessionSlice = createSlice({
     name: 'session',
     initialState,
     reducers: {
-        changeView(state, action: PayloadAction<View>) {
-            state.activeView = action.payload;
-        },
-
         showContextMenu(state, action: PayloadAction<Point>) {
             state.menuContext = {
                 type: MenuType.contextMenu,
